Exports HttpMethod and Route types from wasp/client
