+++ waspc/data/Generator/templates/sdk/client/index.ts
@@ -9,3 +9,15 @@ import { join as joinPaths } from 'path';
 export function resolveProjectPath(path: string): string {
   return joinPaths('../../../', path);
 }
+
+// PUBLIC API
+// NOTE: This is enough to cover Operations and our APIs (src/Wasp/AppSpec/Api.hs).
+export enum HttpMethod {
+	Get = 'GET',
+	Post = 'POST',
+	Put = 'PUT',
+	Delete = 'DELETE',
+}
+
+// PUBLIC API
+export type Route = { method: HttpMethod; path: string }
