--- waspc/data/Generator/templates/sdk/types/index.ts
@@ -1,9 +1,0 @@
-// NOTE: This is enough to cover Operations and our APIs (src/Wasp/AppSpec/Api.hs).
-export enum HttpMethod {
-	Get = 'GET',
-	Post = 'POST',
-	Put = 'PUT',
-	Delete = 'DELETE',
-}
-
-export type Route = { method: HttpMethod; path: string }
